statio: bugfix for checking stats data length
It seems pretty clear this should have always been checking "dlen
& 7U" rather than "dlen & 4U".  I've added a safety against
massively oversized inputs as well, which was already present in
the equivalent Zig version of this code (which already happened to
not have the length checking bug here anyways).

This bug could cause an overflow in the else clause if presented
with a dlen value like "2", and thus then trigger an illegal
memory read and crash the new daemon on startup. To trigger this,
the daemon would have to be attempting control socket takeover
from a supposed other daemon that was, in fact, some other
malicious/broken software.  The control socket stuff has never
claimed to be very secure anyways, but this bug is still worth
fixing.

Thanks to Ryan Steinmetz for reporting this!